+++ blogos-service-impl/src/main/java/com/duan/blogos/service/dao/blog/BlogLikeDao.java
@@ -22,7 +22,7 @@ public interface BlogLikeDao extends BaseDao<BlogLike> {
      * @param blogId    博文id
      * @return 操作影响的行数
      */
-    int deleteLikeByBloggerId(@Param("bloggerId") int bloggerId, @Param("blogId") int blogId);
+    int deleteLikeByBloggerId(@Param("bloggerId") Long bloggerId, @Param("blogId") Long blogId);
 
     /**
      * 根据博文id获得所有喜欢记录
@@ -30,7 +30,7 @@ public interface BlogLikeDao extends BaseDao<BlogLike> {
      * @param blogId 博文id
      * @return 查询结果
      */
-    List<BlogLike> listAllLikeByBlogId(int blogId);
+    List<BlogLike> listAllLikeByBlogId(Long blogId);
 
     /**
      * 统计指定博主给出喜欢的次数
@@ -38,7 +38,7 @@ public interface BlogLikeDao extends BaseDao<BlogLike> {
      * @param bloggerId 博主id
      * @return 数量
      */
-    Integer countLikeByLikerId(int bloggerId);
+    Integer countLikeByLikerId(Long bloggerId);
 
     /**
      * 根据博主id和博文id获取记录
@@ -47,7 +47,7 @@ public interface BlogLikeDao extends BaseDao<BlogLike> {
      * @param blogId    博文id
      * @return 查询记录
      */
-    BlogLike getLike(@Param("bloggerId") int bloggerId, @Param("blogId") int blogId);
+    BlogLike getLike(@Param("bloggerId") Long bloggerId, @Param("blogId") Long blogId);
 
     /**
      * 查询博文
@@ -58,7 +58,7 @@ public interface BlogLikeDao extends BaseDao<BlogLike> {
      * @return 查询结果
      */
     List<BlogLike> listLikeBlog(
-            @Param("bloggerId") int bloggerId,
+            @Param("bloggerId") Long bloggerId,
             @Param("offset") int offset,
             @Param("rows") int rows);
 
@@ -68,5 +68,5 @@ public interface BlogLikeDao extends BaseDao<BlogLike> {
      * @param bloggerId 博主 id
      * @return 只查询博文 id
      */
-    List<BlogLike> listAllIdByBloggerId(int bloggerId);
+    List<BlogLike> listAllIdByBloggerId(Long bloggerId);
 }
